apps/consoomer/pyproject.toml
[tool.poetry]
name = "consoomer"
version = "0.1.0"
description = ""
authors = ["Parker Wahle <regulad@regulad.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
prisma = "^0.14.0"
python-dotenv = "^1.0.1"
aio-pika = "^9.4.2"
redis = { extras = ["hiredis"], version = "^5.0.8" }

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"
pylint = "^3.2.6"
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pylint.format]
max-line-length = "120"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
